+++ src/main/kotlin/gomushin/backend/member/facade/MemberInfoFacade.kt
@@ -3,7 +3,7 @@ package gomushin.backend.member.facade
 import gomushin.backend.alarm.dto.SaveAlarmMessage
 import gomushin.backend.alarm.service.StatusAlarmService
 import gomushin.backend.core.CustomUserDetails
-import gomushin.backend.core.configuration.redis.RedisService
+import gomushin.backend.alarm.service.NotificationRedisService
 import gomushin.backend.couple.domain.service.CoupleInfoService
 import gomushin.backend.member.domain.service.MemberService
 import gomushin.backend.member.domain.service.NotificationService
@@ -24,7 +24,7 @@ class MemberInfoFacade(
     private val notificationService: NotificationService,
     private val statusAlarmService: StatusAlarmService,
     private val coupleInfoService: CoupleInfoService,
-    private val redisService: RedisService
+    private val notificationRedisService: NotificationRedisService
 ) {
     fun getMemberInfo(customUserDetails: CustomUserDetails): MyInfoResponse {
         val member = memberService.getById(customUserDetails.getId())
@@ -71,6 +71,6 @@ class MemberInfoFacade(
     }
 
     fun getMyReceivedNotification(customUserDetails: CustomUserDetails, recentDays: Long): List<SaveAlarmMessage> {
-        return redisService.getAlarms(customUserDetails.getId(), recentDays)
+        return notificationRedisService.getAlarms(customUserDetails.getId(), recentDays)
     }
 }
