+++ src/test/kotlin/gomushin/backend/member/facade/MemberInfoFacadeTest.kt
@@ -2,7 +2,7 @@ package gomushin.backend.member.facade
 
 import gomushin.backend.alarm.service.StatusAlarmService
 import gomushin.backend.core.CustomUserDetails
-import gomushin.backend.core.configuration.redis.RedisService
+import gomushin.backend.alarm.service.NotificationRedisService
 import gomushin.backend.couple.domain.service.CoupleInfoService
 import gomushin.backend.member.domain.entity.Member
 import gomushin.backend.member.domain.entity.Notification
@@ -37,7 +37,7 @@ class MemberInfoFacadeTest {
     @Mock
     private lateinit var coupleInfoService: CoupleInfoService
     @Mock
-    private lateinit var redisService: RedisService
+    private lateinit var notificationRedisService: NotificationRedisService
 
     @InjectMocks
     private lateinit var memberInfoFacade: MemberInfoFacade
